backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
exceptiongroup==1.2.0
Faker==22.4.0
iniconfig==2.0.0
Markdown==3.5.2
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
psycopg2==2.9.9
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
